add review suggestions and fixes #150
