+++ Makefile
@@ -44,4 +44,4 @@ testserver:
 	docker-compose up
 
 ci:
-	go test -mod=vendor -count=1 -v -cover -covermode=count -coverprofile=coverage.out ./...
+	go test -mod=vendor -count=1 -v -cover -race -covermode=count -coverprofile=coverage.out ./...
